% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_baseline_intensities}
\alias{compute_baseline_intensities}
\title{Compute the Baseline Intensity}
\usage{
compute_baseline_intensities(covariates, beta0, alpha, link = exp)
}
\arguments{
\item{covariates}{Covariate information, its exact format is described in the
Details of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{beta0}{Vector of length q, the true parameter of the baseline
intensity.}

\item{alpha}{Non-negative vector of length p (number of vertices), where each
entry gives the individual activity of the corresponding vertex.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}
}
\value{
\code{compute_baseline_intensities} returns a matrix wit p (number of
vertices) rows and the number of columns equals the length of
\code{covariates$times}. The i-th column of the output gives the baseline
intensities of all vertices at the time \code{covariates$times[i]}.
}
\description{
When provided with the covariates, \code{compute_baseline_intensities} returns the
baseline intensities at the times when the covariates change.
}
