% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_omega}
\alias{compute_omega}
\title{Compute the tuning parameter omega according to theory}
\usage{
compute_omega(hawkes, p, T, A, Cg, alpha3, gamma_bar, mu = log(2))
}
\arguments{
\item{p}{The number of vertices in the network}

\item{T}{The end of the observation period}
}
\value{
\code{compute_omega} returns a vector of length \code{p} that contains the
penalty parameter for each vertex. It can be, e.g., provided to
estimate_hawkes() as \code{omega}.
}
\description{
\code{compute_omega} computes a vector of tuning parameters according to the
theoretic results provided in our paper.
}
