% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_observation_matrix}
\alias{create_observation_matrix}
\title{Compute Observation Matrix}
\usage{
create_observation_matrix(n)
}
\arguments{
\item{n}{Integer, the dimension of the required matrix.}
}
\value{
The observation matrix of the required dimension.
}
\description{
The observation matrix of dimension n required to handle the missing
intercept in the model is computed. For details please see our paper.
}
\details{
The usage of this function is explained in the documentations for
\code{\link[=estimate_hawkes]{estimate_hawkes()}} and \code{\link[=debias_Hawkes]{debias_Hawkes()}}
}
