% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_hawkes}
\alias{estimate_hawkes}
\title{Fits a Network Hawkes Model with Covariates}
\usage{
estimate_hawkes(
  covariates,
  hawkes,
  omega,
  omega_alpha,
  lb,
  ub,
  fit_theta = TRUE,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  beta_init = NULL,
  gamma_init = NULL,
  alpha_init = NULL,
  link = exp,
  observation_matrix = NULL
)
}
\arguments{
\item{covariates}{Covariate information, its exact format is described in the
Details of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{hawkes}{The observed point process. This is a list with at least the
element \code{EL}. This element needs to have the same matrix form as in the
output of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{omega}{Vector of length p (number of vertices), containing non-negative
tuning parameter for the LASSO penalty for the corresponding row of the
adjacency matrix.}

\item{omega_alpha}{A non-negative number, containing the tuning parameter for
the LASSO penalty with respect to the individuals intensities a.}

\item{lb, ub}{Vectors of length q+1, q equals the dimension of covariates. The
first q entries of \code{lb} and \code{ub} provide lower and upper bounds on beta,
respectively. The last entry provides a lower (resp. upper) bound on gamma.}

\item{fit_theta}{Logical value, if TRUE (the default) the parameters beta and
gamma are also fitted. If FALSE, beta and gamma are fixed equal to the
provided values in \code{beta_init}, \code{gamma_init}.}

\item{print.level}{Integer which specifies how much information about the
iteration should be printed: 0 (the default) means no information, 1
provides some information and any number larger than 1 results in printing
of all intermediate steps. Note that the nloptr print.level is always equal
to 0.}

\item{max_iteration}{Maximal number of iterations after which the iteration
is stopped. Default is 100.}

\item{tol}{When the parameters have changed after one iteration less than the
value provided in \code{tol}, the iteration is stopped and the current value is
returned as the result. The default is 0.00001.}

\item{beta_init, gamma_init}{If \code{fit_theta=FALSE} these parameters provide
starting values for beta and gamma, respectively. If \code{NULL} (the default),
the middle between \code{lb} and \code{ub} is chosen. If \code{fit_theta=FALSE}, see
\code{fit_theta}. In this case both \code{beta_init} and \code{gamma_init} must be
specified.}

\item{alpha_init}{Vector of initial values for alpha, unless it is \code{NULL}
(the default) in which case all alphas are initialized with 1.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{observation_matrix}{If \code{NULL} (the default), the observation matrix is
computed within \code{estimate_hawkes}. For a single call of \code{estimate_hawkes}
this makes no difference. However, for repeated calls it might be more
efficient to compute the matrix using \code{\link[=create_observation_matrix]{create_observation_matrix()}} and
pass it as a parameter here.}
}
\value{
Returns a list with the elements \code{C}, \code{alpha}, \code{beta}, and \code{gamma}
which contain the estimates for the respective parameters.
}
\description{
\code{estimate_hawkes} estimates the parameter of the Hawkes Causal Model as
described in the paper. It requires a covariate process, cf.
\code{\link[=simulate_hawkes]{simulate_hawkes()}} for a description, the LASSO parameters and lower and
upper bounds on the parameters as inputs. The observed point process is
provided as a list in the same format as returned by \code{\link[=simulate_hawkes]{simulate_hawkes()}}
}
\details{
The estimation is an iterative procedure as mentioned in the paper. In each
iteration, we first update \code{C}, then \code{alpha} and then \code{beta} and \code{gamma}.
Therefore, initial values are required for all parameters other than \code{C}, cf.
the corresponding parameters of the function. These iterations are repeated
until the maximum number of iterations is reached or until the change of the
parameters after one iteration lies below the threshold. It is possible to
provide values for \code{beta} and \code{gamma} and to estimate only \code{C} and \code{alpha}.
}
