% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MultiHawkes_robust}
\alias{MultiHawkes_robust}
\title{Estimate the parameters from a multiple Hawkes process}
\usage{
MultiHawkes_robust(
  multi_covariates,
  multi_hawkes,
  omega,
  omega_alpha,
  lb,
  ub,
  K,
  starting_beta = NULL,
  starting_gamma = NULL,
  C.ind.pen = NULL,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  alpha_init = NULL,
  link = exp,
  observation_matrix = NULL,
  cluster = NULL
)
}
\arguments{
\item{multi_covariates}{A list each entry of which contains a \code{covariate} as
described in, e.g., \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. Each entry of the list corresponds to
another realisation of the Hawkes process.}

\item{multi_hawkes}{A list each entry of which contains a Hawkes process in
the same format as for, e.g, \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. Each entry of
\code{multi_hawkes} corresponds to one realisation of the Hawkes process.}

\item{omega}{Vector of length p (number of vertices), containing non-negative
tuning parameter for the LASSO penalty for the corresponding row of the
adjacency matrix.}

\item{omega_alpha}{A non-negative number, containing the tuning parameter for
the LASSO penalty with respect to the individuals intensities a.}

\item{lb, ub}{Vectors of length q+1, q equals the dimension of covariates. The
first q entries of \code{lb} and \code{ub} provide lower and upper bounds on beta,
respectively. The last entry provides a lower (resp. upper) bound on gamma.}

\item{K}{Number of starting values to use}

\item{starting_beta}{Matrix, each row of which corresponds to a starting
value for beta to be used}

\item{starting_gamma}{Vector, each entry of which corresponds to a starting
value for gamma to be used}

\item{C.ind.pen}{Individual weights for the LASSO estimation of C, which are
passed to glmnet as penalty.factor. The weights are internally rescaled. To
achieve the criterion function from the paper, set it equal to \code{NULL} (the
default), then, all vertices receive the same weight.}

\item{print.level}{Passed to \code{\link[=estimate_hawkes]{estimate_hawkes()}}. If positive, in addition,
information about in which stage the estimation is, is printed. The default
is 0.}

\item{max_iteration}{Maximal number of iterations after which the iteration
is stopped. Default is 100.}

\item{tol}{When the parameters have changed after one iteration less than the
value provided in \code{tol}, the iteration is stopped and the current value is
returned as the result. The default is 0.00001.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{cluster}{If \code{NULL} (the default) serial computations are executed. If a
cluster as returned by \code{makeCluster} is provided (after calling
\code{registerDoParallel(cluster)}), the estimation of C is executed in
parallel. This requires the packages \code{parallel}, \code{doParallel}, and
\code{foreach}.}
}
\value{
\code{MultiHawkes_robust} returns a list containing the estimates for the
parameters \code{C}, \code{alpha}, \code{beta}, and \code{gamma}.
}
\description{
\code{MultiHawkes_robust} estimates the parameters of a Causal Hawkes process as
described in the paper when several realisations of this process are provided.
}
\details{
Suppose that you have several observation from the same Hawkes causal model
as in our paper, that is, the parameters are the same for all realisations but
the covariates might be potentially different. \code{MultiHawkes_robust} estimates
this parameter by optimizing the following loss function:

Sum over k ( Least squares loss for realisation k) + Penalty

where the penalty equals

Sum over i( omega\link{i}\emph{sum(C\link{i,})) + omega_alpha}sum(alpha).

The estimation method is the same as in \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. De-biasing is
not supported for multi observations.
}
