% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiHawkes_surrogate_test}
\alias{multiHawkes_surrogate_test}
\title{Compute edge-wise surrogate permutation test}
\usage{
multiHawkes_surrogate_test(
  multi_hawkes,
  multi_covariates,
  per,
  C0,
  beta0,
  gamma0,
  omega,
  omega_alpha
)
}
\arguments{
\item{multi_covariates}{A list of the same format as in
\code{\link[=MultiHawkes]{MultiHawkes()}}. However, it is required that the \code{times} vector in
each entry is the same.}

\item{per}{Number of permutations in the permuation test, defaults to 100}

\item{C0}{The estimated network against which the shuffled datasets are
compared}

\item{beta0, gamma0}{Parameters to be used for estimating the influences,
these can be, e.g., estimates returned by \code{\link[=MultiHawkes]{MultiHawkes()}}}
}
\value{
A matrix of p-values
}
\description{
Computes a p-value for each edge that qunatifies its significance in the
data. A low p-value indicates that the corresponding edge corrensponds to a
significant influence.
}
\details{
The test is based on a permutation strategy. Specifically, for \code{per} many
repetitions, we reshuffle the events within each instance of the multiple
Hawkes process \code{multiHawkes} and compute a new estimate for the influence
network \code{C} based on the reshuffled data. Because of the reshuffling, the
relation between the processes is removed, i.e., the reshuffled data
corresponds to a dataset with the same marginal frequencies of events, but
there are no mutual influences.
}
