% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NetHawkes_gridsearch}
\alias{NetHawkes_gridsearch}
\title{Compute a Complete Estimator Using Grid Search}
\usage{
NetHawkes_gridsearch(
  covariates,
  hawkes,
  omega,
  omega_alpha,
  lb,
  ub,
  delta = 0.5,
  C.ind.pen = NULL,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  link = exp,
  observation_matrix_network = NULL,
  observation_matrix_debiasing = NULL,
  cluster = NULL,
  debias_thresh = 1e-14,
  debias_maxit = 1e+08,
  debias_exact = TRUE
)
}
\arguments{
\item{covariates}{Covariate information, its exact format is described in the
Details of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{hawkes}{The observed point process. This is a list with at least the
element \code{EL}. This element needs to have the same matrix form as in the
output of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{omega}{Vector of length p (number of vertices), containing non-negative
tuning parameter for the LASSO penalty for the corresponding row of the
adjacency matrix.}

\item{omega_alpha}{A non-negative number, containing the tuning parameter for
the LASSO penalty with respect to the individuals intensities a.}

\item{lb, ub}{Vectors of length q+1, q equals the dimension of covariates. The
first q entries of \code{lb} and \code{ub} provide lower and upper bounds on beta,
respectively. The last entry provides a lower (resp. upper) bound on gamma.}

\item{delta}{Mesh size of the grid to be used. The default is 0.5.}

\item{C.ind.pen}{Individual weights for the LASSO estimation of C, which are
passed to glmnet as penalty.factor. The weights are internally rescaled. To
achieve the criterion function from the paper, set it equal to \code{NULL} (the
default), then, all vertices receive the same weight.}

\item{print.level}{Passed to \code{\link[=estimate_hawkes]{estimate_hawkes()}}. If positive, in addition,
information about in which stage the estimation is, is printed. The default
is 0.}

\item{max_iteration}{Maximal number of iterations after which the iteration
is stopped. Default is 100.}

\item{tol}{When the parameters have changed after one iteration less than the
value provided in \code{tol}, the iteration is stopped and the current value is
returned as the result. The default is 0.00001.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{observation_matrix_network, observation_matrix_debiasing}{Similarly as in
\code{\link[=estimate_hawkes]{estimate_hawkes()}} these matrices are automatically computed when \code{NULL} is
provided here (the default). Since the matrix is the same in repeated calls,
it can lead to a speed up to compute the matrices once using
\code{\link[=create_observation_matrix]{create_observation_matrix()}}. The \code{observation_matrix_network} is of
dimension \code{p} and the \code{observation_matrix_debiasing} is of dimension
\code{1+q+p+p^2-1}. Here \code{p} denotes the number of vertices and \code{q} the dimension
of the covariates.}

\item{cluster}{If \code{NULL} (the default) serial computations are executed. If a
cluster as returned by \code{makeCluster} is provided (after calling
\code{registerDoParallel(cluster)}), the estimation of C is executed in
parallel. This requires the packages \code{parallel}, \code{doParallel}, and
\code{foreach}.}

\item{debias_thresh}{Passed to glmnet(), the default value is \code{1e-14},
glmnet() suggests \code{1e-07}. Changing this value can speed up the
computations.}

\item{debias_maxit}{Passed to glmnet(), the default value is \code{1e+08},
glmnet() suggests \code{1e+05}. Changing this value can speed up the
computations.}

\item{debias_exact}{If \code{TRUE}, the default, the node-wise lasso is computed
exactly by coef() from the glmnet package. If \code{FALSE}, an approximation is
used, which can speed-up computations.}
}
\value{
The retuned value is a list of the same structure as for
\code{\link[=NetHawkes]{NetHawkes()}} but with an two additional elements: \code{nloptr} that contains
the complete output of the refinement call from \code{nloptr}. This allows,
e.g., to check for convergence of the optimization. Moreover, \code{grid}
contains the parameters (\code{beta} first and then \code{gamma}) and the values of
the objective function that were found in the grid search.
}
\description{
\code{NetHawkes_gridsearch} works similarly as \code{\link[=NetHawkes]{NetHawkes()}} but it uses a
different optimization strategy. See Details for a precise description of the
differences.
}
\details{
\code{NetHawkes_gridsearch} uses a grid search (grid of mesh \code{delta}) to determine
the starting value for the optimisation in the first stage: We simply compute
the objective function on a grid of mesh \code{delta}. The minimizing value will
be used as starting value for a subsequent call of the \code{nloptr} package to
find the first stage estimator. The second and third stage remain the same as
in \code{\link[=NetHawkes]{NetHawkes()}}
}
