% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{node_wise_influence_test}
\alias{node_wise_influence_test}
\title{Compute node-wise covariance permutation test}
\usage{
node_wise_influence_test(
  multi_hawkes,
  multi_covariates,
  beta,
  gamma,
  alpha,
  C,
  per = 100,
  link = exp,
  observation_matrix = NULL,
  print.level = 0,
  cluster = NULL
)
}
\arguments{
\item{multi_covariates}{A list of the same format as in
\code{\link[=MultiHawkes]{MultiHawkes()}}. However, it is required that the \code{times} vector in
each entry is the same.}

\item{beta, gamma, alpha, C}{Parameters to be used in the null simulation step,
these can be, e.g., estimates returned by
\code{\link[=MultiHawkes]{MultiHawkes()}}}

\item{per}{Number of permutations in the permuation test, defaults to 100}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{print.level}{Passed to \code{\link[=estimate_hawkes]{estimate_hawkes()}}. If positive, in addition,
information about in which stage the estimation is, is printed. The default
is 0.}

\item{cluster}{If \code{NULL} (the default) serial computations are executed. If a
cluster as returned by \code{makeCluster} is provided (after calling
\code{registerDoParallel(cluster)}), the estimation of C is executed in
parallel. This requires the packages \code{parallel}, \code{doParallel}, and
\code{foreach}.}
}
\value{
A list containing the following three elements:
\itemize{
\item \code{pvals}: The computed p-values per vertex
\item \code{null_test}: A matrix that contains in each of its \code{per} rows the
computed value of the covariance test statistic computed in
the corresponding null simulation
\item \code{cov_test}: The value of the covariance test for the provided dataset
}
}
\description{
Tests for isolated vertices and returns a p-value per vertex. A low p-value
indicates that the corresponding vertex has at least one incoming edge.
}
\details{
The test is based on a permutation strategy. Specifically, to test for
incoming edges in vertex i, \code{per} many datasets are simulated using the
provided parameters, but where the i-th row of \code{C} is replaced by zeros. In
each dataset (including the original dataset), the covariance test statistic
is computed for the LASSO regression that estimates the corresponding row of
the influence matrix \code{C}. The p-values are computed as the fraction of test
statistics in the simulated data that turn out to be larger than the test
statistics of the corresponding vertex in the observed dataset.
}
