% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_count_intensities}
\alias{plot_count_intensities}
\title{Plot Intensity Functions and Hawkes Processes}
\usage{
plot_count_intensities(hawkes, T)
}
\arguments{
\item{hawkes}{A Hawkes process in the form of a list with at least the
elements \code{EL} and \code{intensities} in the form as described in
\code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{T}{Positive number giving the end of observation period.}
}
\value{
\code{plot_count_intensities} generates \code{p} plots (one for each vertex).
The plots show the realized Hawkes processes and the corresponding
intensities. Before calling the function a plot window of the necessary
size has to be created.
}
\description{
\code{plot_count_intensities} plots the provided Hawkes processes along with their
intensity functions. A plot window the provides enough space to plot a graph
for each vertex must be available before calling the function.
}
