% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Visualize Estimated Hawkes Causal Model in a Graph}
\usage{
plot_interactions(
  estHawkes,
  vertex.scaling = 1,
  edge.scaling = 1,
  vertex.names = NULL,
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{estHawkes}{An estimated Hawkes process as returned from
\code{\link[=estimate_hawkes]{estimate_hawkes()}}.}

\item{vertex.scaling, edge.scaling}{Non-negative numbers (1, by default) which
scale edge weights and vertex sizes.}

\item{vertex.names}{Potential names of the vertices in the output graph, can
be \code{NULL} (the default).}

\item{show.plot}{Logical value, if \code{TRUE} (the default) the network is
plotted, otherwise the network is only returned but not plotted.}

\item{...}{Additional arguments passed to the plotting routine of igraph if
\code{show.plot=TRUE}.}
}
\value{
\code{plot_interactions} returns an igraph object with the properties
described under details. If \code{show.plot=TRUE}, the network is in addition
plotted.
}
\description{
An estimate returned via \code{\link[=estimate_hawkes]{estimate_hawkes()}} is visualized as a network.
Moreover, the network is returned. The use of this functions requires the
package igraph.
}
\details{
\code{plot_interactions} generates a weighted network. The edge weights correspond
to the estimated values in \code{C} (potentially scaled by the value of
\code{edge.scaling}). The sizes of the vertices are such that there area is
proportional to \code{(0.1+sqrt(alpha))^2}, where \code{alpha} is the baseline activity
of the corresponding vertex (potentially multiplied with \code{vertex.scaling}).
The vertex names in the graph can be specified through \code{vertex.names}.
}
