% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_hawkes}
\alias{simulate_hawkes}
\title{Simulate an Event Network from provided covariates}
\usage{
simulate_hawkes(
  covariates,
  beta0,
  gamma,
  alpha,
  C,
  T,
  link = exp,
  print.level = 0,
  rand_seed = NA,
  promote_one_event = FALSE,
  beta_poe = 0
)
}
\arguments{
\item{covariates}{Covariate information, its exact format is described in the
Details below.}

\item{beta0}{Vector of length q, the true parameter of the baseline
intensity.}

\item{gamma}{Decay rate (typically positive) in the excitation kernels.}

\item{alpha}{Non-negative vector of length p (number of vertices), where each
entry gives the individual activity of the corresponding vertex.}

\item{C}{Weighted adjacency matrix of the network, must be non-negative and
of dimension p x p.}

\item{T}{Positive number giving the end of observation period.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{print.level}{A single number, if 0 (the default) no status information
is printed, if positive some updates will be given about the process of the
the simulation.}

\item{rand_seed}{Integer, if \code{NA} (the default) the current time stamp is
converted to an integer value. For the simulation it is necessary to
compute random numbers in a C sub-routine, \code{rand_seed} will be passed to
this routine. See also the Details section.}

\item{promote_one_event}{If \code{FALSE} (the default), the simulation is
performed as usual. If \code{TRUE}, another covariate is introduced that
indicates if an event has already occured in the corresponding vertex. The
baseline is then computed with the new covariates and a parameter vector
extended by \code{beta_poe}. This requires \code{link=exp}. See below for details}

\item{beta_poe}{If \code{promote_one_event=FALSE}, this parameter is ignored.
Otherwise, its value extends \code{beta0} to give the effect of an event having
already occured in a process. Its value must be negative to provide correct
results.}
}
\value{
A list with the following elements:
\itemize{
\item \code{EL}: A matrix with four columns, each row corresponds to an event. The
third column specfies in which vertex the event happened and the fourth
column at which time. The rows are ordered according to the fourth column.
The first column contains a unique id (starting from 1) which identifies
the event. The second column specifies from which parent event the current
event was generated, events generated through the baseline have a 0 here.
\item \code{intensities}: A matrix with p rows and each column corresponds to an event
(in the same order as in \code{EL}). Each column gives the intensities of the
corresponding processes at the time of the event.
\item \code{integrals}: A matrix with p rows and each column corresponds to an event
(in the same order as in \code{EL}). Each column provides the values of the
integrals over the excitation kernels with respect to the corresponding
counting process (per vertex) and at the time of the corresponding event.
\item \code{baseline}: A matrix with p rows and number of columns equal to the length
of \code{covariates$times}. The i-th column contains the baseline intensities
of the corresponding vertices at time \code{covariates$times[i]}.
\item \code{covariates}: A copy of the input \code{covariates}.
}
}
\description{
Provides a simulated event history from the Hawkes Causal Model. The true
model parameters and covariate processes have to be provided. In addition to
the simulated events, \code{simulate_hawkes} also returns the overall individual
intensity functions, the baseline intensities and the parts of the intensity
due to mutual excitation.
}
\details{
The result of the function is random and will therefore be different for
different calls unless when the random seed is set before calling the
function and setting the \code{rand_seed} value.

As long as \code{promote_one_event} is set to \code{FALSE}, the provided covariates
will not be altered. If \code{promote_one_event} equals \code{TRUE}, during the
simulation process, a new covariate will be provided, i.e., the existing
covariate matrices will be extended by another column. It will contain a \code{1}
if the corresponding vertex has already experienced an event. The provided
parameter vector \code{beta0} will be extended by \code{beta_poe} and the baseline is
computed using the new sets of covariates and parameter vector. This has the
effect that baseline events occur with a lower probability after the first
event in each process. This feature provides correct results only when
\code{link=exp}. Moreover, only negative values of \code{beta_poe} are handled
correctly.

\code{covariates} is a list of two elements \code{times} and \code{cov}
\itemize{
\item \code{times} is a vector of increasing time points with last element equal to
\code{T}. \code{i}-th element of \code{cov} contains the covariate information valid for
the interval \code{covariates$times[i]} till \code{covariates$times[i+1]}.
\item \code{cov} is itself a list. Its length equals the length of \code{times}. Each
element of \code{cov} is a pxq-matrix, where p is the number of vertices in the
network and q is the number of covariates. Each row of this matrix
corresponds to the covariate vector valid for the respective vertex.
}

Details of the simulation process are provided in our paper.
}
