% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_omega}
\alias{compute_omega}
\title{Compute the tuning parameter omega according to theory}
\usage{
compute_omega(
  hawkes,
  p,
  T,
  gamma_bar,
  mu = log(2),
  alpha3 = 1,
  alpha3_tilde = 0.5,
  N0 = 2,
  Cg = 1
)
}
\arguments{
\item{p}{The number of vertices in the network}

\item{T}{The end of the observation period}

\item{gamma_bar}{The value of gamma to be used in the formula.}

\item{mu, alpha3, alpha3_tilde, N0, Cg}{Additional parameters in the formula.}
}
\value{
\code{compute_omega} returns a vector of length \code{p} that contains the
penalty parameter for each vertex. It can be, e.g., provided to
estimate_hawkes() as \code{omega}.
}
\description{
\code{compute_omega} computes a vector of tuning parameters according to the
theoretic results provided in our paper.
}
