% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cvMultiHawkes}
\alias{cvMultiHawkes}
\title{Use cross-validation to find the tuning parameter for penalizing \code{C}}
\usage{
cvMultiHawkes(
  multi_Hawkes,
  multi_covariates,
  omega_start,
  lb,
  ub,
  nos = 5,
  tf = 0.8,
  M = 5,
  starting_beta = NULL,
  starting_gamma = NULL,
  C.ind.pen = NULL,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  alpha_init = NULL,
  link = exp,
  observation_matrix = NULL,
  cluster = NULL
)
}
\arguments{
\item{multi_covariates}{A list of the same format as in
\code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}}. However, it is required that the \code{times} vector in
each entry is the same.}

\item{omega_start}{A vector (its length must equal the number of vertices)
that contains the starting values of the search. It should contain values
that are too large (see also below for details)}

\item{nos}{The number of starting points to be used in each call of
\code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}}, it defaults to 5.}

\item{tf}{The fraction of the data (in time) that is used for training,
defaults to 0.8.}

\item{M}{The number of steps to be used in the Golden-Section search,
defaults to 5}
}
\value{
A list containing the following three elements:
\itemize{
\item \code{omega}: The optimal choice of the tuning parameter among all evaluated
choices. This can be passed to, e.g, to \code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}}
or \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}.
\item \code{computed_omega}: A matrix with 2M+1 rows. Each row contains a set of
tuning parameters that was evaluated.
\item \code{LSvals}: A matrix with 2M+1 rows. Each row contains the individual
(per vertex) least squares losses of the corresponding tuning
parameters provided in entry \code{computed_omega}.
}
}
\description{
\code{cvMultiHawkes} performs cross validation to find the tuning parameter to be
used for the estimation in \code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}}.
}
\details{
In order to perform cross-validation, the data is split in time in a training
data-set and a testing data-set. The time length of the training data-set is
\code{tf}*\code{T}, where \code{T} is the length of the observation period. The fitting
method is \code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}}. In order to evaluate the performance of a
given tuning parameter, the least squares loss is computed on the remaining
data, the testing data-set. The method to search for the optimal tuning
parameter is Golden-Section search with lower bound 0 and upper bound
provided in \code{omega_start}. While \code{\link[=MultiHawkes_robust]{MultiHawkes_robust()}} is called on the
entire process, the least squares losses are computed for each vertex
separately. Thus, also the optimisation of the tuning parameter is performed
separately.

The Golden-Section search is executed \code{M} times (regardless of convergence of
the algorithm). Finally, the midpoint of the resulting interval is evaluated.
The return value is the choice of the tuning parameter for each vertex that
yields the lowest individual loss.

This routine is also suitable for cross-validation of a single realisation
Hawkes process whenn \code{multi_hawkes} and \code{multi_covariates} contain only one
element.
}
