% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{debias_Hawkes}
\alias{debias_Hawkes}
\title{De-Bias a Given Estimator}
\usage{
debias_Hawkes(
  covariates,
  hawkes,
  est_hawkes,
  link = exp,
  observation_matrix = NULL
)
}
\arguments{
\item{covariates}{Covariate information, its exact format is described in the
Details of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{hawkes}{The observed point process. This is a list with at least the
element \code{EL}. This element needs to have the same matrix form as in the
output of \code{\link[=simulate_hawkes]{simulate_hawkes()}}.}

\item{est_hawkes}{An estimated Hawkes Causal Model, the estimate has to be
formatted as the output of \code{\link[=estimate_hawkes]{estimate_hawkes()}}.}

\item{link}{A function that can be applied to vectors returning a
non-negative vector of the same length, it will be applied to linear
transformation of the covariates with \code{beta0} to obtain the baseline
intensity. The defualt choice is  \code{exp}.}

\item{observation_matrix}{If \code{NULL} (the default), the observation matrix is
computed within \code{estimate_hawkes}. For a single call of \code{estimate_hawkes}
this makes no difference. However, for repeated calls it might be more
efficient to compute the matrix using \code{\link[=create_observation_matrix]{create_observation_matrix()}} and
pass it as a parameter here.}
}
\value{
\code{debias_hawkes} returns a list with the following elements:
\itemize{
\item \code{grad}: A vector containing the derivative of the criterion function.
\item \code{Sigma}: A matrix containing the second derivative of the criterion
function.
\item \code{Theta}: Matrix with q+1 rows (q being the dimension of the covariates)
containing the first q+1 rows of the approximation of the inverse of Sigma
via node-wise LASSO.
\item \code{beta_debiased}: De-biased estimate for beta.
\item \code{gamma_debiased}: De-biased estimate for gamma.
}
}
\description{
\code{debias_hawkes} takes an estimator, e.g., computed through
\code{\link[=estimate_hawkes]{estimate_hawkes()}} and performs the de-biasing on the estimates for \code{beta}
and \code{gamma}.
}
\details{
De-Biasing is required in order to remove the bias from \code{beta} and \code{gamma}
which is introduced in \code{\link[=estimate_hawkes]{estimate_hawkes()}} due to the LASSO penalty.
}
