% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MultiHawkes_robust}
\alias{MultiHawkes_robust}
\title{Estimate the parameters from a multiple Hawkes process}
\usage{
MultiHawkes_robust(
  multi_covariates,
  multi_hawkes,
  omega,
  omega_alpha,
  lb,
  ub,
  K,
  starting_beta = NULL,
  starting_gamma = NULL,
  C.ind.pen = NULL,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  alpha_init = NULL,
  link = exp,
  observation_matrix = NULL,
  cluster = NULL
)
}
\arguments{
\item{multi_covariates}{A list each entry of which contains a \code{covariate} as
described in, e.g., \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. Each entry of the list corresponds to
another realisation of the Hawkes process.}

\item{multi_hawkes}{A list each entry of which contains a Hawkes process in
the same format as for, e.g, \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. Each entry of
\code{multi_hawkes} corresponds to one realisation of the Hawkes process.}
}
\value{
\code{MultiHawkes_robust} returns a list containing the estimates for the
parameters \code{C}, \code{alpha}, \code{beta}, and \code{gamma}.
}
\description{
\code{MultiHawkes_robust} estimates the parameters of a Causal Hawkes process as
described in the paper when several realisations of this process are provided.
}
\details{
Suppose that you have several observation from the same Hawkes causal model
as in our paper, that is, the parameters are the same for all realisations but
the covariates might be potentially different. \code{MultiHawkes_robust} estimates
this parameter by optimizing the following loss function:

Sum over k ( Least squares loss for realisation k) + Penalty

where the penalty equals

Sum over i( omega\link{i}\emph{sum(C\link{i,})) + omega_alpha}sum(alpha).

The estimation method is the same as in \code{\link[=NetHawkes_robust]{NetHawkes_robust()}}. De-biasing is
not supported for multi observations.
}
