% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NetHawkes_robust}
\alias{NetHawkes_robust}
\title{Do not use this function anymore}
\usage{
NetHawkes_robust(
  covariates,
  hawkes,
  omega,
  omega_alpha,
  lb,
  ub,
  C.ind.pen = NULL,
  print.level = 0,
  max_iteration = 100,
  tol = 1e-05,
  link = exp,
  observation_matrix_network = NULL,
  observation_matrix_debiasing = NULL,
  cluster = NULL,
  debias_thresh = 1e-14,
  debias_maxit = 1e+08,
  debias_exact = TRUE
)
}
\description{
Please use \code{\link[=NetHawkes]{NetHawkes()}} instead.
}
