% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{node_wise_influence_test}
\alias{node_wise_influence_test}
\title{Compute node-wise covariance permutation test}
\usage{
node_wise_influence_test(
  multi_hawkes,
  multi_covariates,
  beta,
  gamma,
  alpha,
  C,
  per,
  link = exp,
  observation_matrix = NULL,
  print.level = 0,
  cluster = NULL
)
}
\description{
Compute node-wise covariance permutation test
}
